[config] common 파일에 각 서버 응답 정리
