+++ src/api/store/store.ts
@@ -1,7 +1,8 @@
-import { axiosWithAuth } from 'api/common';
-import { getApiUrl } from 'api/store/common';
+import { axiosWithAuth, getApiUrl } from 'api/common';
+
 import axios from 'axios';
-import type { SuccessOkResponse } from 'api/store/common';
+import type { SuccessOkResponse } from 'api/common';
+import type { StoreDetaillResponse, StoreListResponse } from 'api/store/common';
 
 interface StoreListParams {
   category?: '음식점' | '카페' | '모임';
@@ -16,50 +17,8 @@ interface StoreSearchParams {
   sort?: string;
 }
 
-export interface StoreUser {
-  id: number;
-  name: string;
-  introduction: string;
-  location: string;
-  mainImage: string;
-  open: boolean;
-}
-
-export interface StoreListResponse {
-  curCount: number;
-  curPage: number;
-  totalCount: number;
-  totalPage: number;
-  storeResponseList: StoreUser[];
-}
-
 interface StoreDetaillParams {
   id: number;
-}
-
-export interface StoreDetaillResponse {
-  id: number;
-  name: string;
-  location: string;
-  introduction: string;
-  category: string;
-  telNum: string;
-  mainImage: string | null;
-  monOpenTime: string;
-  monCloseTime: string;
-  tueOpenTime: string;
-  tueCloseTime: string;
-  wedOpenTime: string;
-  webCloseTime: string | null;
-  thuOpenTime: string;
-  thuCloseTime: string;
-  friOpenTime: string;
-  friCloseTime: string;
-  satOpenTime: string | null;
-  satCloseTime: string | null;
-  sunOpenTime: string | null;
-  sunCloseTime: string | null;
-  dayOff: string[];
 }
 
 export class StoreUserApi {
